apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "platform-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "SQLAlchemy==2.0.32",
  "psycopg[binary]==3.2.1",
  "alembic==1.13.2",
  "redis==5.0.8",
  "python-jose==3.3.0",
  "orjson==3.10.7",
]

[tool.setuptools]
packages = ["app"]
